requirements.txt
watchdog
rarfile
yara-python
pefile